counsel-gtags--read-tag: Fix narrowing of candidates
If `global` is called with the completion flag (`-c`) and a string,
the string is always regarded as a prefix. This is not consistent with
the other ivy-commands, where the search string can be anywhere within
a candidate.
To fix this, `global` is only called once and without a prefix string,
which gives us all the symbols available. Any further narrowing will
be handled over to ivy.

I also fixed an incorrect caller value in `counsel-gtags--select-file`.